fix search text and files
